metadata_event_handler/pyproject.toml
[tool.poetry]
name = "metadata_event_handler"
version = "1.0.1"
description = "Consumes Metadata events and writes into ElasticSearch"
authors = ["Indoc Research"]

[tool.poetry.dependencies]
python = "^3.10"
fastavro = "1.5.2"
aiokafka = "0.7.2"
elasticsearch = { extras = ["async"], version = "7.17.4" }
fastapi = "0.88.0"
uvicorn = "^0.18.3"
httpx = "^0.23.0"
pytest-httpx = "^0.21.0"
pydantic = "^1.10.2"
python-dotenv = "0.19.1"
pytest-mock = "^3.8.2"
pilot-platform-common = "0.3.0"

[tool.poetry.dev-dependencies]
Faker = "13.15.1"
pytest = "7.1.2"
pytest-asyncio = "0.18.3"
pytest-cov = "3.0.0"
pytest-random-order = "1.0.4"

[tool.pytest.ini_options]
addopts = "-v --random-order"
testpaths = "tests"
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
